requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
dnslib==0.9.25
dnspython==2.7.0
idna==3.10
requests==2.32.3
scapy==2.6.1
urllib3==2.2.3
